Allow options to contain multiple words
When documenting subcommands, sometimes mkdocs-click will only show the
subcommand name. `:prog_name:` seems to be the right option to ensure
the actuall command is shown, but due to an overly-strict regex, it
used to ignore everything after the first whitespace. With this change,
`prog_name` and other commands can contain multiple words in their
values.

For example

```
::: mkdocs-click
    :module: foo.utils.bridge
    :prog_name: foo-utils bridge
    :command: cli
```

now properly renders as

> foo-utils bridge [OPTIONS]

instead of

> foo-utils [OPTIONS]

or

> bridge [OPTIONS]

(^^ if prog_name wasn't specified at all)

Co-Authored-By: Bar Hochman <bhochman@paloaltonetworks.com>

Fixes #42